move get_precise_epoch to higher level utils
