+++ metaseq/logging/progress_bar/utils.py
@@ -34,11 +34,3 @@ def get_progress_bar(
         bar = WandBProgressBarWrapper(bar, wandb_project, run_name=wandb_run_name)
 
     return bar
-
-
-def get_precise_epoch(epoch: Optional[int], count: int, iterator_size: int) -> float:
-    return (
-        epoch - 1 + (count + 1) / float(iterator_size)
-        if epoch is not None and iterator_size > 0
-        else None
-    )
