+++ metaseq/logging/progress_bar/wandb_progress_bar.py
@@ -5,7 +5,7 @@ from metaseq.logging.progress_bar.base_progress_bar import (
     BaseProgressBar,
     logger,
 )
-from metaseq.logging.progress_bar.utils import get_precise_epoch
+from metaseq.utils import get_precise_epoch
 
 wandb = None
