+++ utils.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+
+def get_precise_epoch(epoch: Optional[int], count: int, iterator_size: int) -> float:
+    return (
+        epoch - 1 + (count + 1) / float(iterator_size)
+        if epoch is not None and iterator_size > 0
+        else None
+    )
